Update main.py
If no french title has been found, we have to set the variable to avoid to get an error when trying to apply newTitleI.lower().

No need to use function to find the quality. CP does it nicely and it gets sometimes an error. You can add it back easily if wanted.

I just want to say, the initial dev did a very nice work ! Thanks